requirements.txt
psycopg2-binary==2.9.9
jupyter==1.1.1
pytest==8.3.5
